+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Handle FCM data notifications (iOS) [#1042](https://github.com/rokwire/illinois-app/issues/1042).
 
 ## [3.0.57] - 2021-11-30
 ### Fixed