+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Handle FCM data notifications (iOS) [#1042](https://github.com/rokwire/illinois-app/issues/1042).
+- Android: release builds [#1046](https://github.com/rokwire/illinois-app/issues/1046).
 
 ## [3.0.57] - 2021-11-30
 ### Fixed