+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Android: release builds [#1046](https://github.com/rokwire/illinois-app/issues/1046).
 ### Added
 - Implement Inbox group message tap action [#1048](https://github.com/rokwire/illinois-app/issues/1048).
+- Hook Notifications Delete user API [#1050](https://github.com/rokwire/illinois-app/issues/1050).
 
 ## [3.0.57] - 2021-11-30
 ### Fixed